consistent naming for utils and utilising shared folder
