+++ packages/artifact/src/internal/find/get-artifact.ts
@@ -6,7 +6,7 @@ import {defaults as defaultGitHubOptions} from '@actions/github/lib/utils'
 import {getRetryOptions} from './retry-options'
 import {requestLog} from '@octokit/plugin-request-log'
 import {GetArtifactResponse} from '../shared/interfaces'
-import {getBackendIdsFromToken} from '../shared/util'
+import {getBackendIdsFromToken} from '../shared/utils'
 import {getUserAgentString} from '../shared/user-agent'
 import {internalArtifactTwirpClient} from '../shared/artifact-twirp-client'
 import {ListArtifactsRequest, StringValue, Timestamp} from '../../generated'
