+++ packages/artifact/src/internal/shared/utils.ts
