+++ packages/cache/__tests__/cacheUtils.test.ts
@@ -1,6 +1,6 @@
 import {promises as fs} from 'fs'
 import * as path from 'path'
-import * as cacheUtils from '../src/internal/cacheUtils'
+import * as cacheUtils from '../src/internal/shared/utils'
 
 beforeEach(() => {
   jest.resetModules()
