+++ packages/cache/__tests__/saveCacheV2.test.ts
@@ -1,7 +1,7 @@
 import * as core from '@actions/core'
 import * as path from 'path'
 import {saveCache} from '../src/cache'
-import * as cacheUtils from '../src/internal/cacheUtils'
+import * as cacheUtils from '../src/internal/shared/utils'
 import {CacheFilename, CompressionMethod} from '../src/internal/constants'
 import * as config from '../src/internal/config'
 import * as tar from '../src/internal/tar'
