+++ packages/cache/src/cache.ts
@@ -1,6 +1,6 @@
 import * as core from '@actions/core'
 import * as path from 'path'
-import * as utils from './internal/cacheUtils'
+import * as utils from './internal/shared/utils'
 import * as cacheHttpClient from './internal/cacheHttpClient'
 import * as cacheTwirpClient from './internal/shared/cacheTwirpClient'
 import {getCacheServiceVersion, isGhes} from './internal/config'
@@ -12,7 +12,7 @@ import {
   FinalizeCacheEntryUploadResponse,
   GetCacheEntryDownloadURLRequest
 } from './generated/results/api/v1/cache'
-import {CacheFileSizeLimit} from './internal/constants'
+import {CacheFileSizeLimit} from './internal/shared/constants'
 export class ValidationError extends Error {
   constructor(message: string) {
     super(message)
