+++ packages/io/src/io.ts
@@ -1,6 +1,6 @@
 import {ok} from 'assert'
 import * as path from 'path'
-import * as ioUtil from './io-util'
+import * as ioUtil from './io-utils'
 
 /**
  * Interface for cp/mv options
